Create WordCopyAssessment
VBA for buttons within an assessment tool. The VBA buttons checked the assessment for omissions and common errors, automatically named the document according to agency guidelines, drafted an email and sent the document to a supervisor through Outlook, and/or copied information from the assessment tool to other required forms.